+++ batched_gemm/cublas_mod.F90
@@ -58,7 +58,7 @@ INTERFACE
         USE ISO_C_BINDING
         CHARACTER(1,C_CHAR), VALUE            :: CTA, CTB
         INTEGER(C_INT),      VALUE            :: M, N, K, LDA, LDB, LDC, TDA, TDB, TDC, BATCHCOUNT
-        REAL(2),             VALUE            :: ALPHA, BETA
+        REAL(4),             VALUE            :: ALPHA, BETA
         REAL(2),             DIMENSION(LDA,*) :: A
         REAL(2),             DIMENSION(LDB,*) :: B
         REAL(4),             DIMENSION(LDC,*) :: C
